cli/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "Envoice CLI Agent Trainer"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "playwright>=1.42.0",
    "pydantic-settings>=2.2.0",
    "jsonschema>=4.21.0",
    "rich>=13.7.0",
    "supabase>=2.4.0",
    "python-dotenv>=1.0.1",
    "jsonpatch>=1.33",
]

[project.scripts]
agent = "agent.main:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 88
target-version = "py39"
